skills/0xnagato/antigravity-swarm/requirements.txt
rich>=13.0.0
pyyaml>=6.0
